+++ jest.config.js
@@ -1,4 +1,10 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  moduleNameMapper: {
+    '\\.css$': 'identity-obj-proxy',
+  },
+  testPathIgnorePatterns: [
+    '\\.tsx$'
+  ],
 };
